use int8 for Bounce field
Reduces model struct size from 544 to 536 bytes. Tis is significant because
we retain this struct in memory.

+++ internal/models/event_model.go
@@ -26,7 +26,7 @@ type Model struct {
 	Duration         int64
 	Id               uint64
 	City             uint32
-	Bounce           int32
+	Bounce           int8
 	View             bool
 	Session          bool
 }
